+++ src/pages/sign-in/index.tsx
@@ -1,0 +1,2 @@
+import { SignIn } from "@/scenes/SignIn";
+export default SignIn;
