--- src/pages/signin/index.tsx
@@ -1,2 +1,0 @@
-import { Signin } from "@/scenes/Signin";
-export default Signin;
